Make Escrow.sol pure ERC20 compatible contract
Some networks(like Polygon Mainnet) have a standard ERC20 implementation of the HMT. This makes it impossible to use HMT native bulk* methods in Escrow contracts. So that's why Escrow's bulkPayOut was modified and now includes the bulk transfer within the contract, instead of executing it on behalf HMT contract.